reomved logic from gameboard and add it into app comp
