+++ 02-tic-tac-toe-starting-project/src/App.jsx
@@ -2,7 +2,15 @@ import { useState } from "react"
 import GameBoard from "./components/GameBoard"
 import Player from "./components/Player"
 import Log from "./components/Log"
+import { WINNING_COMBINATIONS } from './winning-combinations'
 
+
+const initialGameBoard = [
+  [null, null, null],
+  [null, null, null],
+  [null, null, null],
+];
+// helper fuction to get the active game player
 
 function derivedAcitvePlayer(gameTurns) {
   let currentPlayer = 'X'
@@ -16,8 +24,20 @@ function derivedAcitvePlayer(gameTurns) {
 function App() {
 
   const [gameTurns, setGameTurns] = useState([])
-
   const activePlayer = derivedAcitvePlayer(gameTurns)
+
+
+  let gameBoard = initialGameBoard
+
+  for (const turn of gameTurns) {
+    const { square, player } = turn
+    const { row, col } = square
+    gameBoard[row][col] = player
+  }
+
+  for (const combination of WINNING_COMBINATIONS) {
+
+  }
 
   const handleSelectSquare = (rowIndex, colIndex) => {
     // setActivePlayer(preveActivePlayer => preveActivePlayer === "X" ? "O" : "X")
@@ -39,7 +59,7 @@ function App() {
           <Player initialName="Player 1" symbol="O" isAcitve={activePlayer === 'O'} />
 
         </ol>
-        <GameBoard onSelectSquare={handleSelectSquare} activePlayerSymbol={activePlayer} turns={gameTurns} />
+        <GameBoard onSelectSquare={handleSelectSquare} activePlayerSymbol={activePlayer} board={gameBoard} />
       </div>
 
       <Log turns={gameTurns} />
