+++ 02-tic-tac-toe-starting-project/src/components/GameBoard.jsx
@@ -1,18 +1,7 @@
-import React, { useState } from "react";
-const initialGameBoard = [
-    [null, null, null],
-    [null, null, null],
-    [null, null, null],
-];
-function GameBoard({ onSelectSquare, turns }) {
+import React from "react";
 
-    let gameBoard = initialGameBoard
+function GameBoard({ onSelectSquare, board }) {
 
-    for (const turn of turns) {
-        const { square, player } = turn
-        const { row, col } = square
-        gameBoard[row][col] = player
-    }
     // const [gameBoard, setGameBoard] = useState(initialGameBoard)
     // const handleSelectSquare = (rowIndex, colIndex) => {
     //     setGameBoard((prevGameBoard) => {
@@ -25,12 +14,12 @@ function GameBoard({ onSelectSquare, turns }) {
     // }
     return (
         <ol id="game-board">
-            {gameBoard.map((row, rowIndex) => (
+            {board.map((row, rowIndex) => (
                 <li key={rowIndex}>
                     <ol>
                         {row.map((playerSymbol, colIndex) => (
                             <li key={colIndex}>
-                                <button onClick={() => onSelectSquare(rowIndex, colIndex)}>{playerSymbol}</button>
+                                <button onClick={() => onSelectSquare(rowIndex, colIndex)} disabled={playerSymbol !== null}>{playerSymbol}</button>
                             </li>
                         ))}
                     </ol>
